refact move endpoints to api.rs
